+++ src/components/Contact.tsx
@@ -1,0 +1,256 @@
+import emailjs, { EmailJSResponseStatus } from '@emailjs/browser';
+import { useState } from 'react';
+import { BiPhoneCall } from 'react-icons/bi';
+import { FiAlertTriangle } from 'react-icons/fi';
+import { IoMdCheckmark } from 'react-icons/io';
+import { LuMapPin, LuSun } from 'react-icons/lu';
+import { MdAlternateEmail } from 'react-icons/md';
+import contactUsBG from '../assets/images/background/contact-us-bg.svg';
+import SectionTitle from './SectionTitle';
+
+const MOCK_DATA = [
+	{
+		title: 'Email',
+		description: 'mohammedtaysser983@gmail.com',
+		icon: MdAlternateEmail,
+	},
+	{ title: 'Phone', description: '+201015081861', icon: BiPhoneCall },
+	{
+		title: 'Address',
+		description: 'Egypt, Ad Daqahliyah, Markaz Meet Ghamr',
+		icon: LuMapPin,
+	},
+	{ title: 'Working hours', description: '8 a.m. - 10 p.m.', icon: LuSun },
+];
+
+function Contact() {
+	const [isSuccess, setIsSuccess] = useState(false);
+	const [error, setError] = useState<EmailJSResponseStatus | null | string>(
+		null
+	);
+	const [isLoading, setIsLoading] = useState(false);
+	const [formInfo, setFormInfo] = useState({
+		name: '',
+		email: '',
+		message: '',
+		phone: '',
+		subject: '',
+	});
+
+	const onInputChange = (
+		evt: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
+	) => {
+		setFormInfo({
+			...formInfo,
+			[evt.target.name]: evt.target.value,
+		});
+	};
+
+	const onFormSubmit = async (evt: React.FormEvent<HTMLFormElement>) => {
+		evt.preventDefault();
+
+		setIsLoading(true);
+
+		setIsSuccess(false);
+		setError(null);
+
+		const templateParams = {
+			from_name: formInfo.name,
+			to_name: 'Mohammed Taysser',
+			phone: formInfo.phone,
+			email: formInfo.email,
+			message: formInfo.message,
+			subject: formInfo.subject,
+			source: window.location.href,
+		};
+
+		const SERVICE_ID = import.meta.env.VITE_EMAILJS_SERVICE_ID;
+		const TEMPLATE_ID = import.meta.env.VITE_EMAILJS_TEMPLATE_ID;
+		const PUBLIC_KEY = import.meta.env.VITE_EMAILJS_PUBLIC_KEY;
+
+		if (!SERVICE_ID || !TEMPLATE_ID || !PUBLIC_KEY) {
+			setError('Missing env variables');
+			return;
+		}
+
+		await emailjs
+			.send(SERVICE_ID, TEMPLATE_ID, templateParams, {
+				publicKey: PUBLIC_KEY,
+			})
+			.then(
+				function () {
+					setIsSuccess(true);
+
+					setFormInfo({
+						name: '',
+						email: '',
+						message: '',
+						phone: '',
+						subject: '',
+					});
+
+					setIsLoading(false);
+				},
+				function (error) {
+					setError(error);
+				}
+			);
+	};
+
+	return (
+		<section className='contact-me my-5' id='contact-me'>
+			<div className='container my-5 '>
+				<SectionTitle center title='Contact Me' subtitle='Thanks' />
+
+				<div className='pt-5'>
+					<div className='nice-shadow rounded-4 p-1 my-5'>
+						<div className='row align-items-stretch'>
+							<div className='col-md-5'>
+								<div
+									className='contact-information h-100'
+									style={{
+										backgroundImage: `url(${contactUsBG})`,
+									}}
+								>
+									<h3 className='text-white mb-4'>Contact information</h3>
+
+									{MOCK_DATA.map((item, index) => (
+										<div
+											className='row my-2 text-white align-items-center'
+											key={index}
+										>
+											<div className='col-auto'>
+												<item.icon className='fs-2' />
+											</div>
+
+											<div className='col'>
+												<div className='opacity-75 text-small'>
+													{item.title}
+												</div>
+												<p>{item.description}</p>
+											</div>
+										</div>
+									))}
+								</div>
+							</div>
+
+							<div className='col-md-7'>
+								<form onSubmit={onFormSubmit} className='mx-4 mb-3 h-100'>
+									<h3 className='mb-4 mt-4'>Get in touch</h3>
+
+									<div className='row'>
+										<div className='col-md-6 my-3'>
+											<input
+												placeholder='Your name'
+												name='name'
+												required
+												value={formInfo.name}
+												type='text'
+												className='form-control'
+												onChange={onInputChange}
+											/>
+										</div>
+
+										<div className='col-md-6 my-3'>
+											<input
+												placeholder='Your email'
+												name='email'
+												type='email'
+												value={formInfo.email}
+												onChange={onInputChange}
+												required
+												className='form-control'
+											/>
+										</div>
+
+										<div className='col-md-6 my-3'>
+											<input
+												placeholder='Your subject'
+												name='subject'
+												type='text'
+												value={formInfo.subject}
+												onChange={onInputChange}
+												required
+												className='form-control'
+											/>
+										</div>
+
+										<div className='col-md-6 my-3'>
+											<input
+												placeholder='Your phone'
+												name='phone'
+												type='tel'
+												onChange={onInputChange}
+												required
+												value={formInfo.phone}
+												className='form-control'
+											/>
+										</div>
+
+										<div className='col-12 my-3'>
+											<textarea
+												placeholder='Your message'
+												name='message'
+												required
+												rows={3}
+												value={formInfo.message}
+												onChange={onInputChange}
+												className='form-control'
+											/>
+										</div>
+
+										<div className='col-12 my-3'>
+											{isSuccess && (
+												<div
+													className='alert alert-success d-flex gap-2 align-items-center'
+													role='alert'
+												>
+													<IoMdCheckmark size={20} />
+													<div>
+														Thank you for your message, I will get back to you
+														as soon as possible
+													</div>
+												</div>
+											)}
+
+											{error && (
+												<div
+													className='alert alert-error gap-2 d-flex align-items-center'
+													role='alert'
+												>
+													<FiAlertTriangle size={20} />
+													<div>{JSON.stringify(error)}</div>
+												</div>
+											)}
+											<button
+												className='btn btn-primary btn-lg'
+												disabled={isLoading}
+												type='submit'
+											>
+												{isLoading ? (
+													<>
+														<span
+															className='spinner-border spinner-border-sm'
+															aria-hidden='true'
+														></span>
+														<span role='status' className='mx-2'>
+															Loading...
+														</span>
+													</>
+												) : (
+													<span>Send message</span>
+												)}
+											</button>
+										</div>
+									</div>
+								</form>
+							</div>
+						</div>
+					</div>
+				</div>
+			</div>
+		</section>
+	);
+}
+
+export default Contact;